# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Category:
    properties:
      category_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      parent_category:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  domain.CategoryListResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/domain.Category'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  domain.CreateCategoryRequest:
    properties:
      category_name:
        type: string
      parent_category:
        type: string
      slug:
        type: string
    required:
    - category_name
    - slug
    type: object
...
  domain.Order:
    properties:
      alias_id:
        type: string
      created_at:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.OrderItem'
        type: array
      note:
        type: string
      status:
        $ref: '#/definitions/domain.OrderStatus'
      total_amount:
        type: number
    type: object
  domain.OrderItem:
    properties:
      book_id:
        type: string
      name:
        type: string
      quantity:
        type: integer
      unit_price:
        type: number
    type: object
  domain.OrderListResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/domain.Order'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  domain.OrderStatus:
    enum:
    - pending
    - confirmed
    - packing
    - shipping
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusConfirmed
    - OrderStatusPacking
    - OrderStatusShipping
    - OrderStatusCompleted
    - OrderStatusCancelled
  domain.OrderStatusHistory:
    properties:
      alias_id:
        type: string
      changed_at:
        type: string
      changed_by_admin_alias_id:
        type: string
      new_status:
        type: string
      note:
        type: string
      old_status:
        type: string
    type: object
...
  domain.UpdateCategoryRequest:
    properties:
      category_name:
        type: string
      parent_category:
        type: string
      slug:
        type: string
    type: object
  domain.UpdateOrderStatusRequest:
    properties:
      note:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/domain.OrderStatus'
        enum:
        - pending
        - confirmed
        - packing
        - shipping
        - completed
        - cancelled
    required:
    - status
    type: object
...
  server.errorResponse:
    properties:
      error:
        type: string
    type: object
  server.successResponse:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@bookstore.local
    name: API Support
  description: Polyglot persistence bookstore backend — PostgreSQL · MongoDB · Neo4j
    · Redis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: HCMUS Bookstore API
  version: "2.0"
paths:
...
  /admin/categories:
    get:
      description: Return a paginated list of categories for admin management
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CategoryListResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: List categories'
      tags:
      - admin-categories
    post:
      consumes:
      - application/json
      description: Create a new category in MongoDB and sync to Neo4j graph
      parameters:
      - description: Category payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Category'
      security:
      - BearerAuth: []
      summary: 'Admin: Create category'
      tags:
      - admin-categories
  /admin/categories/{id}:
    delete:
      description: Delete a category from MongoDB and detach from Neo4j graph
      parameters:
      - description: Category MongoDB ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: Delete category'
      tags:
      - admin-categories
    put:
      consumes:
      - application/json
      description: Update category metadata and re-sync Neo4j graph
      parameters:
      - description: Category MongoDB ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Category'
      security:
      - BearerAuth: []
      summary: 'Admin: Update category'
      tags:
      - admin-categories
  /admin/orders:
    get:
      description: Return a paginated list of all orders with optional status filter
      parameters:
      - description: Filter by order status
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.OrderListResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: List orders'
      tags:
      - admin-orders
  /admin/orders/{id}:
    get:
      description: Return full order details for any order
      parameters:
      - description: Order Alias ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Order'
      security:
      - BearerAuth: []
      summary: 'Admin: Get order'
      tags:
      - admin-orders
  /admin/orders/{id}/history:
    get:
      description: Return the full history of status transitions for an order
      parameters:
      - description: Order Alias ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.OrderStatusHistory'
            type: array
      security:
      - BearerAuth: []
      summary: 'Admin: Get order audit trail'
      tags:
      - admin-orders
  /admin/orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Transition an order to a new state and log audit trail
      parameters:
      - description: Order Alias ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Status update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.errorResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: Update order status'
      tags:
      - admin-orders
...
  /categories:
    get:
      description: Return a paginated list of all categories
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CategoryListResponse'
      summary: Get categories
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"